+++ docs/source/changelog.rst
@@ -18,6 +18,7 @@ Fixed
 ~~~~~
 * `Issue #10 <https://github.com/pmacg/py-sgtl/issues/10>`_ - correct ``graph.weight`` calculation when edges have floating point weights or self-loops
 * `Issue #3 <https://github.com/pmacg/py-sgtl/issues/3>`_ - correct number of edges for weighted graphs
+* `Issue #8 <https://github.com/pmacg/py-sgtl/issues/8>`_ - return a meaningful error when computing the conductance or bipartiteness of the empty set
 
 0.3.3 - 2021-11-12
 ------------------